Split ConnectApplicationModal into modal and content
