+++ core-ui/src/components/Applications/ConnectApplicationModal/ConnectApplication.scss
